Implement DFS algorithm for network connectivity
Implemented a DFS (Depth-First Search) algorithm in 'dfs_search.py' to verify whether a network is fully connected. The algorithm starts with a random node, checks its neighbors, visits the unvisited ones and continues this process recursively.

The 'search_guide.md' was also updated to mark the completion of tasks related to the implementation of the DFS algorithm.

Additionally, the method for reading JSON data was modified in 'network_object.py' for consistency in file reading across different modules.