Implement flooding search algorithm
Added flooding_search.py file which contains the implementation of the flooding search algorithm using 'NetworkMessage' class. This algorithm is used to broadcast a message to all nodes in a network and then start searching for a specific resource from a given node. It aids in simulating a networked environment where messages are passed between nodes.

Also, the class name 'Message' in network_message.py has been changed to 'NetworkMessage' for better readability and understanding of its function in the system.