Refactor network search method to BFS algorithm
The network search method was refactored from a resource specific search to a generalized use of the Breadth-First Search (BFS) algorithm. This allows flexibility for searches in the network independent of the specific resources being sought, increasing scalability and efficiency. This concise BFS is used to traverse all nodes in the network, starting from a specific node. Along with this, 'search_pipeline' function was also updated to accommodate the BFS search. Also, the 'dfs_search.py' has been renamed to 'fundamental_searches.py', combining both the BFS and DFS methods. The README file was updated to reflect these changes.