Increase TTL, update search target and modify network data
Increased the initial time-to-live (TTL) value in `flooding_search.py` from 2 to 3 to broaden the search and improve chances of finding the desired resource. Modified the 'desiredResource' parameter in `visualizations.py` from 'dancing_moon.mp3' to 'immortal_land.mp3' bearing in mind a change in search target. Added new nodes and made changes to current edges on 'json_example.json' to reflect structural alterations of the network. These modifications are crucial for maintaining up-to-date and realistic network simulations.