Refactored codebase for advanced network visualization
The commit includes several improvements and refactoring of the codebase. Firstly, the in-house html file for visualization was removed and replaced with a Flask-compatible file in the templates directory. This facilitates an easier integration with Flask rendering.

Moreover, redundant code has been removed from the main_app.py file, and a more direct call to generate_network_graph_html method from visualizations.py is used instead.

The generate_network_graph_html function was also added to visualizations.py. This function returns an HTML representation of the network graph, improving the function's reusability and compatibility with Flask's render_template function.

Lastly, Flask version was added to the requirements.txt file ensuring application is developed on the latest and stable version of Flask.