Update import paths and add main_pipeline module
Changed the imports from 'json_files.json_read' to 'json_parsing.json_read' in multiple files to reflect the renaming of the 'json_files' directory to 'json_parsing'. This change was necessary to align the directory name more closely with its purpose of parsing JSON files.

In addition, created a new 'main_pipeline.py' file in the 'wrapper' directory. This module contains functions to manage the search and parsing of JSON files, contributing to the separation of concerns and enhancing maintainability.

Also, minor modifications were made to 'visualizations.py' to adapt function signatures and remove unused code.