- Add backtracking functionality and visual enhancement for path in network graph
Backtracking list functionality added in `flooding_search.py` to keep track of steps taken to find the target node. This function is then used in `visualizations.py` along with an additional color `PATH_NODE_COLOR` for highlighting the backtrack path in the network graph.

This change allows for a better visual understanding of the algorithm's path when searching for a target node. It also changes the json example in `generate_network_graph_html` for further testing and diversity of data representation.